Add files via upload
I have created a Docker task where the local host website asks for a number using the url: localhost:5000/guess/(your_number), if a number is entered wrong, it will display a message of 'Incorrect, try again. Whereas if the guess was correct then it will display 'Correct! You guessed the number.'